Refactor transaction form state handling
uncontrolled -> controlled inputs
DOM state -> Redux state

+++ src/redux/actioncreators.ts
@@ -1,9 +1,9 @@
-import { v4 as uuidv4 } from 'uuid';
-import iobject from '../lang/iobject';
 import DbManager from '../db/manager'
 import { AllState, Services } from '..';
-import { Transaction, Info, ActionMap, DocumentType } from '../types';
+import { Transaction, Info, ActionMap, DocumentType, TransactionFormState, TransactionFormSharedState } from '../types';
 import { ThunkAction } from 'redux-thunk';
+import { createFormData, mapFormDataToTransaction } from '../util/transactionform';
+import selectors from './selectors';
 
 export const CHECK_REMOTE_TAB = 'CHECK_REMOTE_TAB';
 export const CHECK_REMOTE_TAB_FAILURE = 'CHECK_REMOTE_TAB_FAILURE';
@@ -12,6 +12,14 @@ export const IMPORT_TAB = 'IMPORT_TAB';
 export const UPDATE_FROM_DB = 'UPDATE_FROM_DB';
 export const CREATE_OR_UPDATE_TRANSACTION = 'CREATE_OR_UPDATE_TRANSACTION';
 export const REMOVE_TRANSACTION = 'REMOVE_TRANSACTION';
+export const SET_TRANSACTION_FORM = 'SET_TRANSACTION_FORM';
+export const RESET_TRANSACTION_FORM = 'RESET_TRANSACTION_FORM';
+export const UPDATE_TRANSACTION_FORM = 'UPDATE_TRANSACTION_FORM';
+export const UPDATE_TRANSACTION_SHARED_FORM = 'UPDATE_TRANSACTION_SHARED_FORM';
+export const UPDATE_TRANSACTION_DIRECT_FORM = 'UPDATE_TRANSACTION_DIRECT_FORM';
+export const UPDATE_TRANSACTION_PARTICIPANT = 'UPDATE_TRANSACTION_PARTICIPANT';
+export const ADD_PARTICIPANT_TO_TRANSACTION_SHARED_FORM = 'ADD_PARTICIPANT_TO_TRANSACTION_SHARED_FORM';
+export const SET_ALL_JOINED_ON_TRANSACTION_SHARED_FORM = 'SET_ALL_JOINED_ON_TRANSACTION_SHARED_FORM';
 export const SET_ERROR = 'SET_ERROR';
 export const ROUTE_TAB = 'ROUTE_TAB';
 export const ROUTE_TABS = 'ROUTE_TABS';
@@ -86,13 +94,97 @@ const createRemoveTransactionAction = (doc: Transaction): RemoveTransactionActio
   doc
 });
 
+interface SetTransactionFormAction {
+  type: typeof SET_TRANSACTION_FORM;
+  payload: TransactionFormState;
+}
+
+const createSetTransactionFormAction = (payload: TransactionFormState): SetTransactionFormAction => ({
+  type: SET_TRANSACTION_FORM,
+  payload
+});
+
+interface ResetTransactionFormAction {
+  type: typeof RESET_TRANSACTION_FORM;
+}
+
+export const resetTransactionForm = (): ResetTransactionFormAction => ({
+  type: RESET_TRANSACTION_FORM
+});
+
+interface UpdateTransactionFormAction<K extends keyof TransactionFormState> {
+  type: typeof UPDATE_TRANSACTION_FORM;
+  key: K;
+  value: TransactionFormState[K];
+}
+
+export const updateTransactionForm = <K extends keyof TransactionFormState>(key: K, value: TransactionFormState[K]): UpdateTransactionFormAction<K> => ({
+  type: UPDATE_TRANSACTION_FORM,
+  key,
+  value,
+});
+
+interface UpdateTransactionSharedFormAction<K extends keyof TransactionFormState['shared']> {
+  type: typeof UPDATE_TRANSACTION_SHARED_FORM;
+  key: K;
+  value: TransactionFormState['shared'][K];
+}
+
+export const updateTransactionSharedForm = <K extends keyof TransactionFormState['shared']>(key: K, value: TransactionFormState['shared'][K]): UpdateTransactionSharedFormAction<K> => ({
+  type: UPDATE_TRANSACTION_SHARED_FORM,
+  key,
+  value,
+});
+
+interface UpdateTransactionDirectFormAction<K extends keyof TransactionFormState['direct']> {
+  type: typeof UPDATE_TRANSACTION_DIRECT_FORM;
+  key: K;
+  value: TransactionFormState['direct'][K];
+}
+
+export const updateTransactionDirectForm = <K extends keyof TransactionFormState['direct']>(key: K, value: TransactionFormState['direct'][K]): UpdateTransactionDirectFormAction<K> => ({
+  type: UPDATE_TRANSACTION_DIRECT_FORM,
+  key,
+  value,
+});
+
+interface UpdateTransactionParticipantAction<K extends 'participant' | 'status' | 'amount'> {
+  type: typeof UPDATE_TRANSACTION_PARTICIPANT;
+  id: string;
+  key: K;
+  value: TransactionFormSharedState[K];
+}
+
+export const updateTransactionParticipant = <K extends 'participant' | 'status' | 'amount'>(id: string, key: K, value: TransactionFormSharedState[K]): UpdateTransactionParticipantAction<K> => ({
+  type: UPDATE_TRANSACTION_PARTICIPANT,
+  id,
+  key,
+  value,
+});
+
+interface AddParticipantToTransactionSharedFormAction {
+  type: typeof ADD_PARTICIPANT_TO_TRANSACTION_SHARED_FORM;
+}
+
+export const addParticipantToTransactionSharedForm = (): AddParticipantToTransactionSharedFormAction => ({
+  type: ADD_PARTICIPANT_TO_TRANSACTION_SHARED_FORM,
+});
+
+interface SetAllJoinedOnTransactionSharedFormAction {
+  type: typeof SET_ALL_JOINED_ON_TRANSACTION_SHARED_FORM;
+}
+
+export const setAllJoinedOnTransactionSharedForm = (): SetAllJoinedOnTransactionSharedFormAction => ({
+  type: SET_ALL_JOINED_ON_TRANSACTION_SHARED_FORM,
+});
+
 interface SetErrorAction {
   type: typeof SET_ERROR;
   error: any;
   info: any;
 }
 
-export const setError = (error: any, info: any) => ({
+export const setError = (error: any, info: any): SetErrorAction => ({
   type: SET_ERROR,
   error,
   info
@@ -143,7 +235,7 @@ export const navigateToAddTransaction = (tabId: string): NavigateToAddTransactio
   }
 });
 
-export type GTAction = CheckRemoteTabAction | CheckRemoteTabFailureAction | CreateTabAction | ImportTabAction | UpdateFromDbAction | CreateOrUpdateTransactionAction | RemoveTransactionAction | SetErrorAction | SelectTabAction | NavigateToTabsAction | NavigateToUpdateTransactionAction | NavigateToAddTransactionAction;
+export type GTAction = CheckRemoteTabAction | CheckRemoteTabFailureAction | CreateTabAction | ImportTabAction | UpdateFromDbAction | CreateOrUpdateTransactionAction | RemoveTransactionAction | SetTransactionFormAction | ResetTransactionFormAction | UpdateTransactionFormAction<keyof TransactionFormState> | UpdateTransactionSharedFormAction<keyof TransactionFormState['shared']> | UpdateTransactionSharedFormAction<keyof TransactionFormState['shared']> | UpdateTransactionDirectFormAction<keyof TransactionFormState['direct']> | UpdateTransactionParticipantAction<'participant' | 'status' | 'amount'> | AddParticipantToTransactionSharedFormAction | SetAllJoinedOnTransactionSharedFormAction | SetErrorAction | SelectTabAction | NavigateToTabsAction | NavigateToUpdateTransactionAction | NavigateToAddTransactionAction;
 
 export type GTThunkAction = ThunkAction<Promise<void>, AllState, Services, GTAction>;
 
@@ -229,35 +321,49 @@ export const importTabFromUrl = (id: string): GTThunkAction => (dispatch, getSta
   return checkTab(dispatch, id, dbManager);
 };
 
-export const addTransaction = (transaction: Partial<Transaction>): GTThunkAction => async (dispatch, getState, { dbManager }) => {
-  const doc = iobject.merge(transaction, {
-    id: uuidv4(),
-    type: DocumentType.TRANSACTION,
-    tabId: getState().location.payload.tabId
-  }) as Transaction;
+export const addOrUpdateTransaction = (): GTThunkAction => async (dispatch, getState, { dbManager }) => {
+  const state = getState();
+  const tabId = state.location.payload.tabId;
+  const transactionId = state.location.payload.transactionId;
+  const formState = state.app.transactionForm;
 
-  dispatch(createCreateOrUpdateTransactionAction(doc));
+  if (!tabId || !formState) {
+    throw new Error();
+  }
 
-  const tabId = getState().location.payload.tabId;
-  dispatch(selectTab(tabId));
-
-  await dbManager.createDoc(doc);
-};
-
-export const updateTransaction = (transaction: Transaction): GTThunkAction => async (dispatch, getState, { dbManager }) => {
+  const transaction = mapFormDataToTransaction(formState, tabId, transactionId);
+  
   dispatch(createCreateOrUpdateTransactionAction(transaction));
 
-  const tabId = getState().location.payload.tabId;
   dispatch(selectTab(tabId));
 
-  await dbManager.updateDoc(transaction);
+  dispatch(resetTransactionForm());
+
+  if (transactionId) {
+    await dbManager.updateDoc(transaction);
+  } else {
+    await dbManager.createDoc(transaction);
+  }
 };
 
-export const removeTransaction = (doc: Transaction): GTThunkAction => async (dispatch, getState, { dbManager }) => {
+export const removeTransaction = (): GTThunkAction => async (dispatch, getState, { dbManager }) => {
+  const state = getState();
+  const id = state.location.payload.transactionId;
+  let doc;
+  if (id) {
+    doc = state.app.docsById[id] as Transaction;
+  }
+
+  if (!id || !doc) {
+    throw new Error();
+  }
+
   dispatch(createRemoveTransactionAction(doc));
 
-  var tabId = getState().location.payload.tabId;
+  const tabId = state.location.payload.tabId;
   dispatch(selectTab(tabId));
+
+  dispatch(resetTransactionForm());
 
   await dbManager.deleteDoc(doc);
 };
@@ -265,4 +371,19 @@ export const removeTransaction = (doc: Transaction): GTThunkAction => async (dis
 export const closeTransaction = (): GTThunkAction => async (dispatch, getState) => {
   const tabId = getState().location.payload.tabId;
   dispatch(selectTab(tabId));
+
+  dispatch(resetTransactionForm());
+};
+
+export const initTransactionForm = (): GTThunkAction => async (dispatch, getState) => {
+  const state = getState();
+  const transactionId = state.location.payload.transactionId;
+  const transaction = state.app.docsById[transactionId] as Transaction;
+  
+  if (transactionId && !transaction) {
+    throw new Error(`No transaction in the store with the ID "${transactionId}"`);
+  }
+  
+  const formState = createFormData(selectors.getAccounts(state), transaction);
+  dispatch(createSetTransactionFormAction(formState));
 };
